Refactor Blackjack's System.in scanner
I wanted to tackle the issue of resource leak errors showing up in VSCode. Normally I would ignore (not suppress) them, but I want to find a "solution". Maybe this one isn't great, but I like the theory of it: declare Scanner as a static final class variable that is referenced to throughout the program.

This tidies up my original idea of a solution which was to pass the one System.in scanner throughout the program, but that would require many changes throughout the function headers, and I didn't want to do that.